+++ exporter/awscloudwatchlogsexporter/exporter_test.go
@@ -15,13 +15,42 @@
 package awscloudwatchlogsexporter
 
 import (
+	"context"
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/cloudwatchlogs"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/model/pdata"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/cwlogs"
 )
+
+type mockPusher struct {
+	mock.Mock
+}
+
+func (p *mockPusher) AddLogEntry(logEvent *cwlogs.Event) error {
+	args := p.Called(nil)
+	errorStr := args.String(0)
+	if errorStr != "" {
+		return awserr.NewRequestFailure(nil, 400, "").(error)
+	}
+	return nil
+}
+
+func (p *mockPusher) ForceFlush() error {
+	args := p.Called(nil)
+	errorStr := args.String(0)
+	if errorStr != "" {
+		return awserr.NewRequestFailure(nil, 400, "").(error)
+	}
+	return nil
+}
 
 func TestLogToCWLog(t *testing.T) {
 	tests := []struct {
@@ -199,4 +228,41 @@ func TestAttrValue(t *testing.T) {
 			assert.Equal(t, tt.want, got)
 		})
 	}
+}
+
+func TestConsumeLogs(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	factory := NewFactory()
+	expCfg := factory.CreateDefaultConfig().(*Config)
+	expCfg.Region = "us-west-2"
+	expCfg.LogGroupName = "testGroup"
+	expCfg.LogStreamName = "testStream"
+	expCfg.MaxRetries = 0
+	exp, err := newCwLogsPusher(expCfg, componenttest.NewNopExporterCreateSettings())
+	assert.Nil(t, err)
+	assert.NotNil(t, exp)
+	ld := pdata.NewLogs()
+	r := ld.ResourceLogs().AppendEmpty()
+	r.Resource().Attributes().UpsertString("hello", "test")
+	logRecords := r.InstrumentationLibraryLogs().AppendEmpty().Logs()
+	logRecords.EnsureCapacity(5)
+	logRecords.AppendEmpty().SetName("test")
+	assert.Equal(t, 1, ld.LogRecordCount())
+
+	logPusher := new(mockPusher)
+	logPusher.On("AddLogEntry", nil).Return("").Once()
+	logPusher.On("ForceFlush", nil).Return("").Twice()
+	exp.(*exporter).pusher = logPusher
+	require.NoError(t, exp.(*exporter).ConsumeLogs(ctx, ld))
+	require.NoError(t, exp.Shutdown(ctx))
+}
+
+func TestNewExporterWithoutRegionErr(t *testing.T) {
+	factory := NewFactory()
+	expCfg := factory.CreateDefaultConfig().(*Config)
+	expCfg.MaxRetries = 0
+	exp, err := newCwLogsExporter(expCfg, componenttest.NewNopExporterCreateSettings())
+	assert.Nil(t, exp)
+	assert.NotNil(t, err)
 }